requirements.txt
ultralytics
Pillow
numpy
opencv-python
labelImg
matplotlib
plotly
seaborn